FDG-11499 rename/relocating data-table files
