+++ src/components/data-preview/data-preview-section-container/data-preview-section-container.spec.js
@@ -15,7 +15,7 @@ import { render } from '@testing-library/react';
 import { RecoilRoot } from 'recoil';
 import DataPreviewSectionContainer from './data-preview-section-container';
 import { DataTableContext } from '../data-preview-context';
-import { contextProps } from '../../data-table/data-table-test-helper';
+import { contextProps } from '../../table-components/table-test-helper';
 import { columnsConstructorData } from '../../dtg-table/data-table-helper';
 import { mockApiData } from '../../dataset-data/test-helper';
 // describe('DataPreviewSectionContainer initial state', () => {
