+++ src/components/table-components/reset-table-container/reset-table-container.module.scss
