+++ src/components/table-components/reset-table-section/reset-table-section.spec.js
