+++ src/models/ITableHeader.ts
@@ -1,6 +1,6 @@
 import { Table } from '@tanstack/react-table';
 
-export interface IDataTableHeader {
+export interface ITableHeader {
   table: Table<Record<string, unknown>>;
   dataTypes: { [key: string]: string };
   resetFilters: boolean;
